fix: harden serve literal-template fallback per review
Review findings on the ${workspaceFolder} tolerance:

- Wrong-project guard: the discarded-template fallback now requires a
  unique registered project when resolution reaches the global registry
  (ServeGlobalDbMatch::UniqueOnly). The cwd depth heuristic only ran for
  genuine no-path serve; a host's spawn directory (typically the user
  home) says nothing about the intended workspace, so silently serving
  the shallower of two different-depth projects risked attaching the
  wrong index. serve also logs which project the fallback picked and
  why.
- Fallible stderr writes: diagnostics use writeln!(io::stderr()) so a
  host that spawned serve with stderr closed cannot panic the server.
- Tighter template detection: brace contents must look like a variable
  name (leading ASCII letter, word/./- chars, optional :-modifier), so
  ${}, ${ }, and ${a/b} are not misclassified as templates.
- Docs: cursor-plugin/README.md is the canonical rationale home; the
  root README, USER-GUIDE, and CHANGELOG carry one-line mentions, and
  "falls back" is qualified with "where possible" (serve still exits on
  ambiguity or an empty registry). The canonical resolution-order doc
  in src/config.rs now describes the template pre-filter step.
- Test structure: template-fallback tests moved to their own
  serve_template_path_test module (wired into the nextest Windows
  group), sharing a serve_harness spawn helper with mcp_cli_serve_test
  instead of duplicating it; new test pins the different-depth
  strictness split between no-path and discarded-template serve.
